+++ Classes/Controller/ContentElements/QuickNavigationContentElementController.php
@@ -1,6 +1,8 @@
 <?php
 namespace T3v\T3vNavigations\Controller\ContentElements;
 
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+
 use T3v\T3vContent\Controller\ContentElements\ContentElementController;
 
 /**
@@ -23,8 +25,18 @@ class QuickNavigationContentElementController extends ContentElementController {
    * @return void
    */
   public function indexAction() {
-    $primaryItems   = $this->quickNavigationItemRepository->findByUids($this->settings['primaryItems']);
-    $secondaryItems = $this->quickNavigationItemRepository->findByUids($this->settings['secondaryItems']);
+    $primaryItems       = [];
+    $primaryItemsUids   = GeneralUtility::intExplode(',', $this->settings['primaryItems'], true);
+    $secondaryItems     = [];
+    $secondaryItemsUids = GeneralUtility::intExplode(',', $this->settings['secondaryItems'], true);
+
+    foreach($primaryItemsUids as $uid) {
+      $primaryItems[] = $this->quickNavigationItemRepository->findByUid($uid);
+    }
+
+    foreach($secondaryItemsUids as $uid) {
+      $secondaryItems[] = $this->quickNavigationItemRepository->findByUid($uid);
+    }
 
     $this->view->assign('primaryItems', $primaryItems);
     $this->view->assign('secondaryItems', $secondaryItems);