+++ Classes/Domain/Model/QuickNavigationItem.php
@@ -18,7 +18,14 @@ class QuickNavigationItem extends AbstractModel {
   protected $title;
 
   /**
-   * The quick navigation item's link.
+   * The quick navigation item's page (UID).
+   *
+   * @var int
+   */
+  protected $page;
+
+  /**
+   * The quick navigation item's link (typolink).
    *
    * @var string
    */
@@ -44,18 +51,37 @@ class QuickNavigationItem extends AbstractModel {
   }
 
   /**
-   * Returns the quick navigation item's link.
+   * Returns the quick navigation item's page (UID).
    *
-   * @return string The quick navigation item's link
+   * @return int The quick navigation item's page (UID)
+   */
+  public function getPage() {
+    return $this->page;
+  }
+
+  /**
+   * Sets the quick navigation item's page (UID).
+   *
+   * @param int $link The quick navigation item's link page (UID)
+   * @return void
+   */
+  public function setPage($page) {
+    $this->page = $page;
+  }
+
+  /**
+   * Returns the quick navigation item's link (typolink).
+   *
+   * @return string The quick navigation item's link (typolink)
    */
   public function getLink() {
     return $this->link;
   }
 
   /**
-   * Sets the quick navigation item's link.
+   * Sets the quick navigation item's link (typolink).
    *
-   * @param string $link The quick navigation item's link
+   * @param string $link The quick navigation item's link (typolink)
    * @return void
    */
   public function setLink($link) {