+++ CHANGELOG.md
@@ -6,8 +6,9 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
-* Implemented `Quick Navigation Item` model
-* Implemented `Overview Navigation Item` model
+* Implemented `Quick Navigation` Content Element
+* Implemented `Quick Navigation Item` Model
+* Implemented `Overview Navigation Item` Model
 * Added common files
 
 [Semantic Versioning]: http://semver.org "Semantic Versioning"